Finish most of UI redesign.

+++ src/project/chain.rs
@@ -5,7 +5,7 @@ pub const ROWS_PER_CHAIN: usize = 16;
 #[derive(Default, Clone, PartialEq, Serialize, Deserialize, Debug)]
 pub struct ChainRow {
     pub phrase: Option<u32>,
-    pub transpose: f32,
+    pub transpose: i32,
 }
 
 #[derive(Clone, PartialEq, Serialize, Deserialize, Debug)]
